fix: Corrects CypherTypecast in expression_tree_walker().
The argument of CypherTypeCast must be passed to "walker". but,
"expression_tree_walker" is called again and the "walker" cannot perform the
required tasks.

So, changed to call walker. If required, "expression_tree_walker" is called
back from "walker".

In addition, this corrects errors caused by Cypher not detecting the nested
aggregate function.